# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A workout tracking web application for personal use (or family, friends), geared towards running and other GPX-based activities",
        "title": "Workout Tracker",
        "contact": {
            "url": "https://github.com/jovandeginste/workout-tracker/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jovandeginste/workout-tracker?tab=License-1-ov-file"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/workouts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "List all workouts of the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/database.Workout"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    }
                }
            }
        },
        "/workouts/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new workout",
                "parameters": [
                    {
                        "description": "Workout data",
                        "name": "workout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.ManualWorkout"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/app.ManualWorkout"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/workouts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get all information about a workout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include details",
                        "name": "details",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/database.Workout"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.APIResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "results": {}
            }
        },
...
        "app.ManualWorkout": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "duration_hours": {
                    "type": "integer"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "repetitions": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/database.WorkoutType"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "database.Equipment": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Whether this equipment is active",
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "default_for": {
                    "description": "Which workout types to add this equipment by default",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.WorkoutType"
                    }
                },
                "description": {
                    "description": "More information about the equipment",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "The name of the gear",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/database.User"
                },
                "userID": {
                    "description": "The ID of the user who owns the workout",
                    "type": "integer"
                },
                "workouts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Workout"
                    }
                }
            }
        },
...
        "database.GPXData": {
            "type": "object",
            "properties": {
                "checksum": {
                    "description": "The checksum of the content",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "content": {
                    "description": "The file content",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "description": "The filename of the file",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workoutID": {
                    "description": "The ID of the workout",
                    "type": "integer"
                }
            }
        },
        "database.MapCenter": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "Latitude",
                    "type": "number"
                },
                "lng": {
                    "description": "Longitude",
                    "type": "number"
                },
                "tz": {
                    "description": "Timezone",
                    "type": "string"
                }
            }
        },
        "database.MapData": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The address of the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/geo.Address"
                        }
                    ]
                },
                "addressString": {
                    "description": "The generic location of the workout",
                    "type": "string"
                },
                "averageSpeed": {
                    "description": "The average speed of the workout",
                    "type": "number"
                },
                "averageSpeedNoPause": {
                    "description": "The average speed of the workout without pausing",
                    "type": "number"
                },
                "center": {
                    "description": "The center of the workout (in coordinates)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.MapCenter"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "description": "The tool that created this workout",
                    "type": "string"
                },
                "details": {
                    "description": "The details of the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.MapDataDetails"
                        }
                    ]
                },
                "extraMetrics": {
                    "description": "Extra metrcis available",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "maxElevation": {
                    "description": "The maximum elevation of the workout",
                    "type": "number"
                },
                "maxSpeed": {
                    "description": "The maximum speed of the workout",
                    "type": "number"
                },
                "minElevation": {
                    "description": "The minimum elevation of the workout",
                    "type": "number"
                },
                "name": {
                    "description": "The name of the workout",
                    "type": "string"
                },
                "pauseDuration": {
                    "description": "The total pause duration of the workout",
                    "type": "integer"
                },
                "start": {
                    "description": "The start time of the workout",
                    "type": "string"
                },
                "stop": {
                    "description": "The stop time of the workout",
                    "type": "string"
                },
                "totalDistance": {
                    "description": "The total distance of the workout",
                    "type": "number"
                },
                "totalDown": {
                    "description": "The total distance down of the workout",
                    "type": "number"
                },
                "totalDuration": {
                    "description": "The total duration of the workout",
                    "type": "integer"
                },
                "totalRepetitions": {
                    "description": "The number of repetitions of the workout",
                    "type": "integer"
                },
                "totalUp": {
                    "description": "The total distance up of the workout",
                    "type": "number"
                },
                "totalWeight": {
                    "description": "The weight of the workout",
                    "type": "number"
                },
                "type": {
                    "description": "The type of the workout",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workoutID": {
                    "description": "The workout this data belongs to",
                    "type": "integer"
                }
            }
        },
        "database.MapDataDetails": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mapDataID": {
                    "description": "The ID of the map data these details belong to",
                    "type": "integer"
                },
                "points": {
                    "description": "The GPS points of the workout",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.MapPoint"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "database.Profile": {
            "type": "object",
            "properties": {
                "api_active": {
                    "description": "Whether the user's API key is active",
                    "type": "boolean"
                },
                "auto_import_directory": {
                    "description": "The user's preferred directory for auto-import",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "description": "The user's preferred language",
                    "type": "string"
                },
                "prefer_full_date": {
                    "description": "Whether to show full dates in the workout details",
                    "type": "boolean"
                },
                "preferredUnits": {
                    "description": "The user's preferred units",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.UserPreferredUnits"
                        }
                    ]
                },
                "socials_disabled": {
                    "description": "Whether social sharing buttons are disabled when viewing a workout",
                    "type": "boolean"
                },
                "theme": {
                    "description": "The user's preferred color scheme",
                    "type": "string"
                },
                "timezone": {
                    "description": "The user's preferred timezone",
                    "type": "string"
                },
                "totals_show": {
                    "description": "What workout type of totals to show",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.WorkoutType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "description": "The ID of the user who owns this profile",
                    "type": "integer"
                }
            }
        },
...
        "database.RouteSegmentMatch": {
            "type": "object",
            "properties": {
                "distance": {
                    "description": "The total distance of the route segment for this workout",
                    "type": "number"
                },
                "duration": {
                    "description": "The total duration of the route segment for this workout",
                    "type": "integer"
                },
                "firstID": {
                    "description": "The index of the first and last point of the route",
                    "type": "integer"
                },
                "lastID": {
                    "description": "The index of the first and last point of the route",
                    "type": "integer"
                },
                "routeSegment": {
                    "$ref": "#/definitions/database.RouteSegment"
                },
                "routeSegmentID": {
                    "description": "The ID of the route segment",
                    "type": "integer"
                },
                "workout": {
                    "$ref": "#/definitions/database.Workout"
                },
                "workoutID": {
                    "description": "The ID of the workout",
                    "type": "integer"
                }
            }
        },
...
        "database.User": {
            "type": "object",
            "properties": {
                "-": {
                    "description": "The salt used to hash the user's password",
                    "type": "string"
                },
                "active": {
                    "description": "Whether the user is active",
                    "type": "boolean"
                },
                "admin": {
                    "description": "Whether the user is an admin",
                    "type": "boolean"
                },
                "apikey": {
                    "description": "The user's API key",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastVersion": {
                    "description": "Which version of the app the user has last seen and acknowledged",
                    "type": "string"
                },
                "name": {
                    "description": "The user's name",
                    "type": "string"
                },
                "profile": {
                    "description": "The user's profile settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.Profile"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "The user's username",
                    "type": "string"
                }
            }
        },
...
        "database.Workout": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "description": "The map data associated with the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.MapData"
                        }
                    ]
                },
                "date": {
                    "description": "The timestamp the workout was recorded",
                    "type": "string"
                },
                "dirty": {
                    "description": "Whether the workout has been modified and the details should be re-rendered",
                    "type": "boolean"
                },
                "equipment": {
                    "description": "Which equipment is used for this workout",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Equipment"
                    }
                },
                "gpx": {
                    "description": "The file data associated with the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.GPXData"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "The name of the workout",
                    "type": "string"
                },
                "notes": {
                    "description": "The notes associated with the workout, in markdown",
                    "type": "string"
                },
                "publicUUID": {
                    "description": "UUID to publicly share a workout - this UUID can be rotated",
                    "type": "string"
                },
                "routeSegmentMatches": {
                    "description": "Which route segments match",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.RouteSegmentMatch"
                    }
                },
                "type": {
                    "description": "The type of the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.WorkoutType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "description": "The user who owns the workout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/database.User"
                        }
                    ]
                },
                "userID": {
                    "description": "The ID of the user who owns the workout",
                    "type": "integer"
                }
            }
        },
...
        "database.WorkoutType": {
            "type": "string",
            "enum": [
                "auto",
                "running",
                "cycling",
                "e-cycling",
                "walking",
                "skiing",
                "snowboarding",
                "swimming",
                "kayaking",
                "golfing",
                "hiking",
                "push-ups",
                "weight-lifting"
            ],
            "x-enum-varnames": [
                "WorkoutTypeAutoDetect",
                "WorkoutTypeRunning",
                "WorkoutTypeCycling",
                "WorkoutTypeECycling",
                "WorkoutTypeWalking",
                "WorkoutTypeSkiing",
                "WorkoutTypeSnowboarding",
                "WorkoutTypeSwimming",
                "WorkoutTypeKayaking",
                "WorkoutTypeGolfing",
                "WorkoutTypeHiking",
                "WorkoutTypePushups",
                "WorkoutTypeWeightLifting"
            ]
        },
        "geo.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "formattedAddress": {
                    "type": "string"
                },
                "houseNumber": {
                    "type": "string"
                },
                "postcode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "stateCode": {
                    "type": "string"
                },
                "stateDistrict": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "suburb": {
                    "type": "string"
                }
            }
        },
...
    }
}